Add recipes and cookbook

+++ src/main/java/edu/ntnu/idi/idatt/model/Recipe.java
@@ -8,7 +8,7 @@ public class Recipe {
   String name;
   String description;
   String instructions;
-  List<Map.Entry<Grocery, Integer>> ingredients;
+  List<Map.Entry<Grocery, Double>> ingredients;
 
   /**
    * Constructor for a recipe.
@@ -22,7 +22,7 @@ public class Recipe {
       String name,
       String description,
       String instructions,
-      List<Map.Entry<Grocery, Integer>> ingredients) {
+      List<Map.Entry<Grocery, Double>> ingredients) {
     this.name = name;
     this.description = description;
     this.instructions = instructions;
@@ -88,7 +88,7 @@ public class Recipe {
    *
    * @return The ingredients needed to make the recipe.
    */
-  public List<Map.Entry<Grocery, Integer>> getIngredients() {
+  public List<Map.Entry<Grocery, Double>> getIngredients() {
     return ingredients;
   }
 
@@ -97,7 +97,7 @@ public class Recipe {
    *
    * @param ingredients The ingredients needed to make the recipe.
    */
-  public void setIngredients(List<Map.Entry<Grocery, Integer>> ingredients) {
+  public void setIngredients(List<Map.Entry<Grocery, Double>> ingredients) {
     this.ingredients = ingredients;
   }
 }
